+++ .firebase/hosting.YnVpbGQ.cache
@@ -1,0 +1,9 @@
+robots.txt,1680607371319,bfe106a3fb878dc83461c86818bf74fc1bdc7f28538ba613cd3e775516ce8b49
+asset-manifest.json,1680866897130,ea58c1250dfe4494247af6e5985ed06eb55d4b0934014a4a65a9f520782a8a40
+favicon.ico,1680866722823,64ee150e2a1fa1a084fd286e9a6d3d9d6e07dfc08cbcb9f1e840c7734135667f
+static/css/main.11dfa7fc.css,1680866897143,7e016181a5eba4430bf1b7845c6de5b64202c6ae30a185fdf4ee0eb7408b059f
+index.html,1680866897129,de8026f4643649a27b4508bd1651df52e171d7591e38e5200021af00d01c11a8
+static/css/main.11dfa7fc.css.map,1680866897145,70f30ff68fc3a10a0e294f75ad7de5222f701505fb91e4c3ec4b68ce8c22be90
+static/js/main.a193cab7.js.LICENSE.txt,1680866897144,dc544f46b2db9e42a6cb6d8644566ecaddbf1a7a071cd2671dacdbf3a285c8dc
+static/js/main.a193cab7.js,1680866897145,acb5d751b978b5119db4ca06ce1fbf19bb03612991fa1b6b4ecdf4e8f42648d2
+static/js/main.a193cab7.js.map,1680866897149,6dc6048038df53e63b03afbbd6176e85f49bc01e5ac5cbb82850eaa153f6bd0b
