+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "wanted-pre-onboarding-fr-39654"
+  }
+}
