+++ public/index.html
@@ -8,11 +8,12 @@
       name="description"
       content="Web site created using create-react-app"
     />
+    <link rel="icon" href="./favicon.ico" />
 
-    <title>React App</title>
+    <title>Todo list</title>
   </head>
   <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <noscript>Just todo list</noscript>
     <div id="root"></div>
   </body>
 </html>
