+++ src/components/Sign-/Sign.module.css
@@ -1,0 +1,25 @@
+.form {
+  margin-left: 10%;
+}
+
+.form button {
+  width: 30%;
+  height: 100%;
+  border: none;
+  border-radius: 9px;
+  cursor: pointer;
+  margin-bottom: 5%;
+  margin-left: 10%;
+  padding: 10px;
+}
+
+.form input {
+  color: black;
+  padding-left: 15px;
+  width: 600px;
+  height: 40px;
+  font-size: 20px;
+  border-radius: 10px;
+  margin-bottom: 7%;
+  outline: none;
+}
