+++ src/components/Todo-/Todo.module.css
@@ -1,0 +1,64 @@
+.form button {
+  width: 50px;
+  height: 100%;
+  border: none;
+  border-radius: 3px;
+  cursor: pointer;
+}
+
+.form input {
+  color: black;
+  padding-left: 15px;
+  width: 600px;
+  height: 100%;
+  font-size: 15px;
+  border-radius: 5px;
+  outline: none;
+}
+
+.ul li {
+  width: 650px;
+  list-style: none;
+  line-height: 40px;
+  position: relative;
+  margin-bottom: 9px;
+  text-indent: 15px;
+  background: rgb(241, 242, 241);
+  border-radius: 3px;
+  font-weight: 400;
+  cursor: pointer;
+}
+
+.ul button {
+  font-weight: 600;
+  text-indent: 15px;
+  display: flex;
+  width: 11.5%;
+  margin-left: 80%;
+}
+
+.ul form {
+  color: black;
+  padding-left: 15px;
+  width: 633px;
+  height: 100%;
+  font-size: 15px;
+  border-radius: 5px;
+  outline: none;
+}
+
+.logout {
+  font-weight: 600;
+  text-indent: 15px;
+  display: flex;
+  width: 100%;
+  margin-top: 30px;
+  justify-content: right;
+  color: inherit;
+  text-decoration: none;
+}
+
+.logout:hover {
+  color: burlywood;
+  text-decoration-line: underline;
+}
